refactor(applist): created a custom component for the application list item (#60)
The application list item was becoming too complex to be just a method inside the application list component.
It had to handle the visualization of the app, open the app edit dialog, open the app delete dialog, etc.
A new component (ApplicationListItem) has been created: this component just manage the visualization of the application and expose
some events. Currently supported events are onEdit and onDelete.